Refactor: Move `convert_to_ruby_type` into file descriptor extension
Similar to ba068d4d320bf26a414071a87b8f481c1657f833, this logic is better on the file descriptor itself rather than a responsibility of the code generator.
